requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
PyYAML==6.0.2
beautifulsoup4==4.12.3
